Refactor depth estimation components: migrate input processing to image_utils and remove InputProcessor class
